Refactor mfenced, msqrt, and mroot into common and chtml components
