+++ mathjax3-ts/output/chtml/Wrappers/mroot.ts
@@ -21,12 +21,11 @@
  * @author dpvc@mathjax.org (Davide Cervone)
  */
 
-import {CHTMLWrapper} from '../Wrapper.js';
+import {CHTMLWrapper, CHTMLConstructor} from '../Wrapper.js';
 import {CHTMLmsqrt} from './msqrt.js';
-import {CHTMLmo} from './mo.js';
+import {CommonMroot, CommonMrootInterface, MrootConstructor} from '../../common/Wrappers/mroot.js';
 import {BBox} from '../BBox.js';
 import {MmlMroot} from '../../../core/MmlTree/MmlNodes/mroot.js';
-import {MmlNode} from '../../../core/MmlTree/MmlNode.js';
 
 /*****************************************************************/
 /**
@@ -36,22 +35,9 @@ import {MmlNode} from '../../../core/MmlTree/MmlNode.js';
  * @template T  The Text node class
  * @template D  The Document class
  */
-export class CHTMLmroot<N, T, D> extends CHTMLmsqrt<N, T, D> {
+export class CHTMLmroot<N, T, D> extends CommonMroot<N, T, D, MrootConstructor>(CHTMLmsqrt) {
+
     public static kind = MmlMroot.prototype.kind;
-
-    /**
-     * @override
-     */
-    get surd() {
-        return 2;
-    }
-
-    /**
-     * @override
-     */
-    get root(): number {
-        return 1;
-    }
 
     /**
      * @override
@@ -65,42 +51,6 @@ export class CHTMLmroot<N, T, D> extends CHTMLmsqrt<N, T, D> {
         if (dx) {
             this.adaptor.setStyle(this.adaptor.firstChild(ROOT) as N, 'paddingLeft', this.em(dx));
         }
-    }
-
-    /**
-     * @override
-     */
-    protected combineRootBBox(BBOX: BBox, sbox: BBox) {
-        const bbox = this.childNodes[this.root].getBBox();
-        const [x, h] = this.getRootDimens(sbox);
-        BBOX.combine(bbox, 0, h);
-    }
-
-    /**
-     * @override
-     */
-    protected getRootDimens(sbox: BBox) {
-        const surd = this.childNodes[this.surd] as CHTMLmo<N, T, D>;
-        const bbox = this.childNodes[this.root].getBBox();
-        const offset = (surd.size < 0 ? .5 : .6) * sbox.w;
-        const {w, rscale} = bbox;
-        const W = Math.max(w, offset / rscale);
-        const dx = Math.max(0, W - w);
-        const h = this.rootHeight(bbox, sbox, surd.size);
-        const x = W * rscale - offset;
-        return [x, h, dx];
-    }
-
-    /**
-     * @param {BBox} rbox      The bbox of the root
-     * @param {BBox} sbox      The bbox of the surd
-     * @param {number} size    The size of the surd
-     * @return {number}        The height of the root within the surd
-     */
-    protected rootHeight(rbox: BBox, sbox: BBox, size: number) {
-        const H = sbox.h + sbox.d;
-        const b = (size < 0 ? 2 + .3 * (H - 4) : .55 * H) - sbox.d;
-        return b + Math.max(0, rbox.d * rbox.rscale);
     }
 
 }
